+++ core/src/main/java/org/elasticsearch/index/analysis/WordDelimiterGraphTokenFilterFactory.java
@@ -98,8 +98,4 @@ public class WordDelimiterGraphTokenFilterFactory extends AbstractTokenFilterFac
         }
         return 0;
     }
-
-    int getFlags() {
-        return flags;
-    }
 }
