+++ docs/reference/analysis/tokenfilters/word-delimiter-graph-tokenfilter.asciidoc
@@ -3,7 +3,7 @@
 
 experimental[]
 
-Named `word_delimiter_graph`, it Splits words into subwords and performs
+Named `word_delimiter_graph`, it splits words into subwords and performs
 optional transformations on subword groups. Words are split into
 subwords with the following rules:
